requirements.txt
altair=4.2.0=pypi_0
attrs=22.1.0=pypi_0
blinker=1.5=pypi_0
ca-certificates=2022.07.19=hecd8cb5_0
cachetools=5.2.0=pypi_0
certifi=2022.9.14=py39hecd8cb5_0
charset-normalizer=2.1.1=pypi_0
click=8.1.3=pypi_0
commonmark=0.9.1=pypi_0
decorator=5.1.1=pypi_0
entrypoints=0.4=pypi_0
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
idna=3.4=pypi_0
importlib-metadata=4.12.0=pypi_0
jinja2=3.1.2=pypi_0
jsonschema=4.16.0=pypi_0
libcxx=14.0.6=h9765a3e_0
libffi=3.3=hb1e8313_2
markupsafe=2.1.1=pypi_0
ncurses=6.3=hca72f7f_3
numpy=1.23.3=pypi_0
openssl=1.1.1q=hca72f7f_0
packaging=21.3=pypi_0
pandas=1.5.0=pypi_0
pillow=9.2.0=pypi_0
pip=22.2.2=py39hecd8cb5_0
protobuf=3.20.3=pypi_0
pyarrow=9.0.0=pypi_0
pydeck=0.8.0b3=pypi_0
pygments=2.13.0=pypi_0
pympler=1.0.1=pypi_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.18.1=pypi_0
python=3.9.13=hdfd78df_1
python-dateutil=2.8.2=pypi_0
pytz=2022.2.1=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
readline=8.1.2=hca72f7f_1
requests=2.28.1=pypi_0
rich=12.5.1=pypi_0
semver=2.13.0=pypi_0
setuptools=63.4.1=py39hecd8cb5_0
six=1.16.0=pypi_0
smmap=5.0.0=pypi_0
sqlite=3.39.3=h707629a_0
streamlit=1.13.0=pypi_0
tk=8.6.12=h5d9f67b_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
tornado=6.2=pypi_0
typing-extensions=4.3.0=pypi_0
tzdata=2022.4=pypi_0
tzlocal=4.2=pypi_0
urllib3=1.26.12=pypi_0
validators=0.20.0=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.6=hca72f7f_0
zipp=3.8.1=pypi_0
zlib=1.2.12=h4dc903c_3
